// src/generated/api/customers/models/adminCustomerSummaryDtoKind.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

/**
 * Suy từ việc khách có tài khoản đăng nhập hay không
 */
export type AdminCustomerSummaryDtoKind =
  (typeof AdminCustomerSummaryDtoKind)[keyof typeof AdminCustomerSummaryDtoKind];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminCustomerSummaryDtoKind = {
  MEMBER: 'MEMBER',
  GUEST: 'GUEST',
} as const;