// src/generated/api/customers/models/adminCustomerSummaryDtoStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type AdminCustomerSummaryDtoStatus =
  (typeof AdminCustomerSummaryDtoStatus)[keyof typeof AdminCustomerSummaryDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminCustomerSummaryDtoStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;